requirements.txt
google-api-python-client==2.137.0
google-auth==2.32.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
python-dotenv==1.0.1
feedparser==6.0.11
openai==1.54.4
httpx==0.27.2